requirements.txt

tensorflow
tensorflow-datasets
jupyterlab

matplotlib

Pillow
numpy
scipy